fix: rename hook useScrollToTop
